test(bdd): migrate runner artifact journeys (phase 3)
