+++ bdd/data/sequences/python-bdd-exception/main.py
@@ -1,2 +1,3 @@
 def main(context, input_stream, *args):
+    print("PYTHON_STDOUT_BEFORE_EXCEPTION", flush=True)
     raise Exception("TestException: This exception should appear on stderr")
