+++ bdd/features/e2e/E2E-014-python.feature
@@ -5,7 +5,11 @@ Feature: Test our shiny new Python runner
         Given host is running
         When find and upload sequence "python-bdd-exception.tar.gz"
         And instance started
-        Then "stderr" contains "TestException: This exception should appear on stderr"
+        And keep instance streams "stdout,stderr"
+        And get runner PID
+        And runner has ended execution
+        Then kept instance stream "stdout" should be "PYTHON_STDOUT_BEFORE_EXCEPTION\n"
+        And "stderr" contains "TestException: This exception should appear on stderr"
         And host is still running
 
     @ci-instance-python
